Level 1 and Level 2 Quiz Preparation/04 OpenAI Agents SDK_Tools/pyproject.toml
[project]
name = "04-openai-agents-sdk-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.3.1",
    "python-dotenv>=1.1.1",
]
